Initial JS interface idea
