+++ src/index.js
@@ -1,6 +1,5 @@
 import WebARKit from './WebARKit'
-import WebARKitController from './WebARKitController'
 
 export default {
-  WebARKit, WebARKitController
+  WebARKit
 }
